+++ ChangeLog.md
@@ -8,6 +8,7 @@
 
     - New TTP::Meteor::isDevel() function, thus bumping minor candidate version number
     - meteor.pl deploy honors '--dummy' option
+    - meteor.pl list --applications has '--dirs' and '--diffs' new options to list the directory when it is doesn't correspond to the name of the application
 
 ### 4.29.0